+++ cmd/warehouses/warehouses.go
@@ -1,9 +1,12 @@
 package warehouses
 
 import (
+	"time"
+
 	query_history "github.com/databricks/bricks/cmd/warehouses/query-history"
 	"github.com/databricks/bricks/lib/sdk"
 	"github.com/databricks/bricks/lib/ui"
+	"github.com/databricks/databricks-sdk-go/retries"
 	"github.com/databricks/databricks-sdk-go/service/warehouses"
 	"github.com/spf13/cobra"
 )
@@ -17,9 +20,14 @@ var Cmd = &cobra.Command{
 }
 
 var createWarehouseReq warehouses.CreateWarehouseRequest
+var createWarehouseAndWait bool
+var createWarehouseTimeout time.Duration
 
 func init() {
 	Cmd.AddCommand(createWarehouseCmd)
+
+	createWarehouseCmd.Flags().BoolVar(&createWarehouseAndWait, "wait", true, `wait to reach RUNNING state`)
+	createWarehouseCmd.Flags().DurationVar(&createWarehouseTimeout, "timeout", 20*time.Minute, `maximum amount of time to reach RUNNING state`)
 	// TODO: short flags
 
 	createWarehouseCmd.Flags().IntVar(&createWarehouseReq.AutoStopMins, "auto-stop-mins", createWarehouseReq.AutoStopMins, `The amount of time in minutes that a SQL Endpoint must be idle (i.e., no RUNNING queries) before it is automatically stopped.`)
@@ -49,6 +57,19 @@ var createWarehouseCmd = &cobra.Command{
 	RunE: func(cmd *cobra.Command, args []string) error {
 		ctx := cmd.Context()
 		w := sdk.WorkspaceClient(ctx)
+		if createWarehouseAndWait {
+			spinner := ui.StartSpinner()
+			info, err := w.Warehouses.CreateWarehouseAndWait(ctx, createWarehouseReq,
+				retries.Timeout[warehouses.GetWarehouseResponse](createWarehouseTimeout),
+				func(i *retries.Info[warehouses.GetWarehouseResponse]) {
+					spinner.Suffix = i.Info.Health.Summary
+				})
+			spinner.Stop()
+			if err != nil {
+				return err
+			}
+			return ui.Render(cmd, info)
+		}
 		response, err := w.Warehouses.CreateWarehouse(ctx, createWarehouseReq)
 		if err != nil {
 			return err
@@ -58,9 +79,14 @@ var createWarehouseCmd = &cobra.Command{
 }
 
 var deleteWarehouseReq warehouses.DeleteWarehouse
+var deleteWarehouseAndWait bool
+var deleteWarehouseTimeout time.Duration
 
 func init() {
 	Cmd.AddCommand(deleteWarehouseCmd)
+
+	deleteWarehouseCmd.Flags().BoolVar(&deleteWarehouseAndWait, "wait", true, `wait to reach DELETED state`)
+	deleteWarehouseCmd.Flags().DurationVar(&deleteWarehouseTimeout, "timeout", 20*time.Minute, `maximum amount of time to reach DELETED state`)
 	// TODO: short flags
 
 	deleteWarehouseCmd.Flags().StringVar(&deleteWarehouseReq.Id, "id", deleteWarehouseReq.Id, `Required.`)
@@ -78,6 +104,19 @@ var deleteWarehouseCmd = &cobra.Command{
 	RunE: func(cmd *cobra.Command, args []string) error {
 		ctx := cmd.Context()
 		w := sdk.WorkspaceClient(ctx)
+		if deleteWarehouseAndWait {
+			spinner := ui.StartSpinner()
+			info, err := w.Warehouses.DeleteWarehouseAndWait(ctx, deleteWarehouseReq,
+				retries.Timeout[warehouses.GetWarehouseResponse](deleteWarehouseTimeout),
+				func(i *retries.Info[warehouses.GetWarehouseResponse]) {
+					spinner.Suffix = i.Info.Health.Summary
+				})
+			spinner.Stop()
+			if err != nil {
+				return err
+			}
+			return ui.Render(cmd, info)
+		}
 		err := w.Warehouses.DeleteWarehouse(ctx, deleteWarehouseReq)
 		if err != nil {
 			return err
@@ -87,9 +126,14 @@ var deleteWarehouseCmd = &cobra.Command{
 }
 
 var editWarehouseReq warehouses.EditWarehouseRequest
+var editWarehouseAndWait bool
+var editWarehouseTimeout time.Duration
 
 func init() {
 	Cmd.AddCommand(editWarehouseCmd)
+
+	editWarehouseCmd.Flags().BoolVar(&editWarehouseAndWait, "wait", true, `wait to reach RUNNING state`)
+	editWarehouseCmd.Flags().DurationVar(&editWarehouseTimeout, "timeout", 20*time.Minute, `maximum amount of time to reach RUNNING state`)
 	// TODO: short flags
 
 	editWarehouseCmd.Flags().IntVar(&editWarehouseReq.AutoStopMins, "auto-stop-mins", editWarehouseReq.AutoStopMins, `The amount of time in minutes that a SQL Endpoint must be idle (i.e., no RUNNING queries) before it is automatically stopped.`)
@@ -121,6 +165,19 @@ var editWarehouseCmd = &cobra.Command{
 	RunE: func(cmd *cobra.Command, args []string) error {
 		ctx := cmd.Context()
 		w := sdk.WorkspaceClient(ctx)
+		if editWarehouseAndWait {
+			spinner := ui.StartSpinner()
+			info, err := w.Warehouses.EditWarehouseAndWait(ctx, editWarehouseReq,
+				retries.Timeout[warehouses.GetWarehouseResponse](editWarehouseTimeout),
+				func(i *retries.Info[warehouses.GetWarehouseResponse]) {
+					spinner.Suffix = i.Info.Health.Summary
+				})
+			spinner.Stop()
+			if err != nil {
+				return err
+			}
+			return ui.Render(cmd, info)
+		}
 		err := w.Warehouses.EditWarehouse(ctx, editWarehouseReq)
 		if err != nil {
 			return err
@@ -130,9 +187,14 @@ var editWarehouseCmd = &cobra.Command{
 }
 
 var getWarehouseReq warehouses.GetWarehouse
+var getWarehouseAndWait bool
+var getWarehouseTimeout time.Duration
 
 func init() {
 	Cmd.AddCommand(getWarehouseCmd)
+
+	getWarehouseCmd.Flags().BoolVar(&getWarehouseAndWait, "wait", true, `wait to reach RUNNING state`)
+	getWarehouseCmd.Flags().DurationVar(&getWarehouseTimeout, "timeout", 20*time.Minute, `maximum amount of time to reach RUNNING state`)
 	// TODO: short flags
 
 	getWarehouseCmd.Flags().StringVar(&getWarehouseReq.Id, "id", getWarehouseReq.Id, `Required.`)
@@ -150,6 +212,19 @@ var getWarehouseCmd = &cobra.Command{
 	RunE: func(cmd *cobra.Command, args []string) error {
 		ctx := cmd.Context()
 		w := sdk.WorkspaceClient(ctx)
+		if getWarehouseAndWait {
+			spinner := ui.StartSpinner()
+			info, err := w.Warehouses.GetWarehouseAndWait(ctx, getWarehouseReq,
+				retries.Timeout[warehouses.GetWarehouseResponse](getWarehouseTimeout),
+				func(i *retries.Info[warehouses.GetWarehouseResponse]) {
+					spinner.Suffix = i.Info.Health.Summary
+				})
+			spinner.Stop()
+			if err != nil {
+				return err
+			}
+			return ui.Render(cmd, info)
+		}
 		response, err := w.Warehouses.GetWarehouse(ctx, getWarehouseReq)
 		if err != nil {
 			return err
@@ -254,9 +329,14 @@ var setWorkspaceWarehouseConfigCmd = &cobra.Command{
 }
 
 var startWarehouseReq warehouses.StartWarehouse
+var startWarehouseAndWait bool
+var startWarehouseTimeout time.Duration
 
 func init() {
 	Cmd.AddCommand(startWarehouseCmd)
+
+	startWarehouseCmd.Flags().BoolVar(&startWarehouseAndWait, "wait", true, `wait to reach RUNNING state`)
+	startWarehouseCmd.Flags().DurationVar(&startWarehouseTimeout, "timeout", 20*time.Minute, `maximum amount of time to reach RUNNING state`)
 	// TODO: short flags
 
 	startWarehouseCmd.Flags().StringVar(&startWarehouseReq.Id, "id", startWarehouseReq.Id, `Required.`)
@@ -274,6 +354,19 @@ var startWarehouseCmd = &cobra.Command{
 	RunE: func(cmd *cobra.Command, args []string) error {
 		ctx := cmd.Context()
 		w := sdk.WorkspaceClient(ctx)
+		if startWarehouseAndWait {
+			spinner := ui.StartSpinner()
+			info, err := w.Warehouses.StartWarehouseAndWait(ctx, startWarehouseReq,
+				retries.Timeout[warehouses.GetWarehouseResponse](startWarehouseTimeout),
+				func(i *retries.Info[warehouses.GetWarehouseResponse]) {
+					spinner.Suffix = i.Info.Health.Summary
+				})
+			spinner.Stop()
+			if err != nil {
+				return err
+			}
+			return ui.Render(cmd, info)
+		}
 		err := w.Warehouses.StartWarehouse(ctx, startWarehouseReq)
 		if err != nil {
 			return err
@@ -283,9 +376,14 @@ var startWarehouseCmd = &cobra.Command{
 }
 
 var stopWarehouseReq warehouses.StopWarehouse
+var stopWarehouseAndWait bool
+var stopWarehouseTimeout time.Duration
 
 func init() {
 	Cmd.AddCommand(stopWarehouseCmd)
+
+	stopWarehouseCmd.Flags().BoolVar(&stopWarehouseAndWait, "wait", true, `wait to reach STOPPED state`)
+	stopWarehouseCmd.Flags().DurationVar(&stopWarehouseTimeout, "timeout", 20*time.Minute, `maximum amount of time to reach STOPPED state`)
 	// TODO: short flags
 
 	stopWarehouseCmd.Flags().StringVar(&stopWarehouseReq.Id, "id", stopWarehouseReq.Id, `Required.`)
@@ -303,6 +401,19 @@ var stopWarehouseCmd = &cobra.Command{
 	RunE: func(cmd *cobra.Command, args []string) error {
 		ctx := cmd.Context()
 		w := sdk.WorkspaceClient(ctx)
+		if stopWarehouseAndWait {
+			spinner := ui.StartSpinner()
+			info, err := w.Warehouses.StopWarehouseAndWait(ctx, stopWarehouseReq,
+				retries.Timeout[warehouses.GetWarehouseResponse](stopWarehouseTimeout),
+				func(i *retries.Info[warehouses.GetWarehouseResponse]) {
+					spinner.Suffix = i.Info.Health.Summary
+				})
+			spinner.Stop()
+			if err != nil {
+				return err
+			}
+			return ui.Render(cmd, info)
+		}
 		err := w.Warehouses.StopWarehouse(ctx, stopWarehouseReq)
 		if err != nil {
 			return err
